+++ src/index.spec.ts
@@ -1,0 +1,7 @@
+import { expose } from './index';
+
+describe('Libary setup', () => {
+  it('should export `expose`', () => {
+    expect(expose).toBeDefined();
+  });
+});